Fix soft abort when Remove Selected button pressed
If the user has only picked an XY point, but not yet a LONLAT point on the OSM layer, when Remove Selected button is pressed and the removeOptions() function is executed, attempting to perform the window.mapOSM.removeLayer(pointLayerOSM) call soft aborts.  Thus the remainder of the logic in the function does not execute, and the symptom is the XY value not getting cleared in the input field.  The added IF statement ensures we do not called the removeLayer unless the pointLayerOSM actually has features.  Tested in Firefox, IE, and Safari.